% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basepeak_finder.R
\name{basepeak_finder}
\alias{basepeak_finder}
\title{Detect basepeak of XXX}
\usage{
basepeak_finder(PuIncR = NULL, XCMSmode = "maxo", XCMSet = NULL,
  ULtag = NULL, Ltag = NULL, separator = "_", sep.pos = NULL,
  UL.atomM = NULL, L.atomM = NULL, ppm.s = NULL, rt.win.min = 1,
  Basepeak.minInt = NULL, Basepeak.Percentage = 0.7, noise.quant = 0.2,
  ...)
}
\arguments{
\item{PuIncR}{Result of PuInc_seeker}

\item{XCMSet}{the xcmsSet with labelled and unlabelled samples.}

\item{ULtag}{part of the samplename for unlabelled sample}

\item{Ltag}{part of the samplename for unlabelled sample}

\item{sep.pos}{unsure.}

\item{UL.atomM}{}

\item{L.atomM}{}

\item{ppm.s}{}

\item{rt.win.min}{}

\item{Basepeak.minInt}{}

\item{Basepeak.Percentage}{}

\item{noise.quant}{}

\item{XCMSmode="maxo"}{Which type of peak intensity to use}

\item{separator="_"}{separator in samplename}
}
\value{
Return basepeak
}
\description{
Longer Description of basepeak_finder
}

