% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_query.R
\name{database_query}
\alias{database_query}
\title{Find Monoisotopic Peaks in a Custom Metabolite Database}
\usage{
database_query(geoRgeR = NULL, adducts = NULL, db = NULL, ppm.db = 10)
}
\arguments{
\item{geoRgeR}{Result of basepeak_finder}

\item{adducts}{List of adducts in the current ion mode}

\item{db}{Database in .CSV format (see \code{system.file("extdata/ExampleDatabase.csv", package="geoRge")})}

\item{ppm.db}{ppm value to search in Database}
}
\description{
Look for hits for mz of monoisotopic peaks in a database in .CSV format
}
