\name{geoRge-package}
\alias{geoRge-package}
\alias{geoRge}
\docType{package}
\title{
\packageTitle{geoRge}
}
\description{
\packageDescription{geoRge}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{geoRge}
\packageIndices{geoRge}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{geoRge}

Maintainer: \packageMaintainer{geoRge}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
