% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_compare.R
\name{label_compare}
\alias{label_compare}
\title{Calculate and Compare Labelling Values between Conditions}
\usage{
label_compare(
  geoRgeR = NULL,
  XCMSet = NULL,
  ppm.s = NULL,
  control.cond = NULL,
  fc.vs.Control = 1,
  p.value.vs.Control = 0.05,
  Show.bp = T
)
}
\arguments{
\item{geoRgeR}{Result of basepeak_finder}

\item{XCMSet}{The xcmsSet with labelled and unlabelled samples}

\item{ppm.s}{ppm window to use to search the monoisotopic peak}

\item{control.cond}{Condition tag to be used as reference for the Welch t.test statistics}

\item{fc.vs.Control}{Default fold-change value to applied to report changes between different conditions}

\item{p.value.vs.Control}{Default p-value to applied to report changes between different conditions}

\item{Show.bp}{Boolean that switches if the monoisotopic labelling percentages should be reported in the output table}
}
\value{
Dataframe with results of comparing the enrichment between \code{control.cond} and the rest of conditions
}
\description{
Calculate the proportion of labelling of the isotopologues found for each group of features and statistically compare them. If only one condition is available only the labelling proportions are reported
}
