% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PuInc_seeker.R
\name{PuInc_seeker}
\alias{PuInc_seeker}
\title{Detect putative Isotope pairs}
\usage{
PuInc_seeker(XCMSet = NULL, XCMSmode = "maxo", ULtag = NULL,
  Ltag = NULL, separator = "_", sep.pos = NULL, fc.threshold = 1.2,
  p.value.threshold = 0.05, PuInc.int.lim = 5000, ...)
}
\arguments{
\item{XCMSet}{the xcmsSet with labelled and unlabelled samples.}

\item{ULtag}{part of the samplename for unlabelled sample}

\item{Ltag}{part of the samplename for unlabelled sample}

\item{sep.pos}{unsure.}

\item{XCMSmode="maxo"}{Which type of peak intensity to use}

\item{separator="_"}{separator in samplename}

\item{fc.threshold=1.2}{Fold-change threshhold between XXX}

\item{p.value.threshold=0.05}{p-Value threshhold between XXX}

\item{PuInc.int.lim=5000}{Some intensity threshhold}
}
\value{
Potential Isotope pairs
}
\description{
Detect putative Isotope pairs
}

