% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepeak_finder.R
\name{refinegeoRge}
\alias{refinegeoRge}
\title{Refine basepeak_finder Results}
\usage{
refinegeoRge(finderR, minAtoms = 3, maxAtoms = 50, minCorr = 0.25)
}
\arguments{
\item{minAtoms}{Minimum number of atoms of an inc_id cluster}

\item{maxAtoms}{Maximum number of atoms of an inc_id cluster}

\item{minCorr}{Minimum \link{stats::cor} correlation of M0 and Mn isotopologues along labelled samples.}

\item{PuIncR}{Result of basepeak_finder}
}
\value{
A refined geoRge data.frame object.
}
\description{
Performs a cleaning step of geoRge results based on number of isotopologues and correlation
}
\note{
Depending on the number of samples, the correlation value may be oversensitive, be careful with this argument.
}
