% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepeak_finder.R
\name{basepeak_finder}
\alias{basepeak_finder}
\title{Find Monoisotopic Peaks for enriched features}
\usage{
basepeak_finder(
  PuIncR = NULL,
  XCMSet = NULL,
  UL.atomM = NULL,
  L.atomM = NULL,
  ppm.s = NULL,
  rt.win.min = 1,
  Basepeak.minInt = NULL,
  Basepeak.Percentage = 0.7,
  noise.quant = 0
)
}
\arguments{
\item{PuIncR}{Result of PuInc_seeker}

\item{XCMSet}{The xcmsSet with labelled and unlabelled samples}

\item{UL.atomM}{Mass of the Unlabelled atom used in labelling experiments}

\item{L.atomM}{Mass of the Labelled atom used in labelling experiments}

\item{ppm.s}{ppm window to use to search the monoisotopic peak}

\item{rt.win.min}{Minimum retention time window in which the isotopologues are expected to coelute}

\item{Basepeak.minInt}{Minimum value of intensity/Area for the candidate to be a candidate monoisotopic peak}

\item{Basepeak.Percentage}{If more than one high-intensity base peak is found, a percentage of the highest one is used to avoid the assignation of high natural abundant peaks belonging to a monoisotopic peak}
}
\value{
A list containing: \describe{
\item{geoRge}{Dataframe containing results for features for which a base peak was found}
\item{params}{Inputed arguments to be inherited by \code{\link{label_compare}}}}
}
\description{
Find the monoisotopic peak of the putative isotopologues detected by PuInc_seeker()
}
