\name{geoRge-package}
\alias{geoRge-package}
\alias{geoRge}
\docType{package}
\title{
\packageTitle{geoRge}
}
\description{
\packageDescription{geoRge}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{geoRge}
\packageIndices{geoRge}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{geoRge}

Maintainer: \packageMaintainer{geoRge}
}
\references{
"geoRge: a computational tool for stable isotope labelling detection in LC/MS-based untargeted metabolomics"

Jordi Capellades, Miriam Navarro, Sara Samino, Marta Garcia-Ramirez, Cristina Hernandez, Rafael Simo, Maria Vinaixa and Oscar Yanes

Analytical Chemistry. 2015.
}
\keyword{ package }
