% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PuInc_seeker.R
\name{PuInc_seeker}
\alias{PuInc_seeker}
\title{Detect Putative Incorporation features}
\usage{
PuInc_seeker(
  XCMSet = NULL,
  XCMSmode = "maxo",
  ULtag = NULL,
  Ltag = NULL,
  separator = "_",
  sep.pos.front = TRUE,
  fc.threshold = 1.2,
  p.value.threshold = 0.05,
  PuInc.int.lim = NULL
)
}
\arguments{
\item{XCMSet}{\emph{xcmsSet} or \emph{XCMSnExp} object of samples processed by xcms workflow}

\item{XCMSmode}{Type of quantification value for the features, as in \code{xcms::groupval(XCMSet)}}

\item{ULtag}{Unlabelled sample name tag, for example "C12"}

\item{Ltag}{Labelled sample name tag, for example "C13"}

\item{separator}{Character used as separator in sample names, as in \code{xcms::sampclass(XCMSet)}}

\item{sep.pos.front}{\code{TRUE} if position of \code{ULtag} and \code{Ltag} is found \emph{previous} to the condition names. Check \code{xcms::sampclass(XCMSet)}}

\item{fc.threshold}{Fold-change threshold for detection of upregulated features associated to Putative incorporation}

\item{p.value.threshold}{p-Value threshhold between for detection of upregulated features associated to Putative incorporation}

\item{PuInc.int.lim}{Intensity limit in UNlabelled sample for the selection of features to be tested as possible Putative incorporation}
}
\value{
A list containing: \describe{
  \item{PuInc}{Putative Incorporation Features matrix}
  \item{PuInc_conditions}{Experimetal conditions (as in \code{xcms::sampclass(XCMSet)}) in which PuInc were detected}
  \item{pvalue}{Matrix of p-values from the tested features}
  \item{foldchange}{Matrix of fold-changes from the tested features}  
  \item{params}{Inputed arguments to be inherited by \code{\link{basepeak_finder}}} }
}
\description{
Use Welch's T-test to detect features upregulated in labelled samples
}
\examples{
data(mtbls213)
xcms::sampclass(mtbls213)
PuInc_seeker(XCMSet=mtbls213,ULtag="CELL_Glc12",Ltag="CELL_Glc13",
sep.pos.front=TRUE ,fc.threshold=1.5,p.value.threshold=.05,PuInc.int.lim = 4000)
}
